signaling-server/requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
Flask-CORS==3.0.10
python-engineio==4.2.1
python-socketio==5.4.0
Werkzeug==2.0.3
# MongoDB dependencies
Flask-PyMongo==2.3.0
pymongo==4.6.2
Flask-Login==0.5.0
Flask-WTF==1.0.0
email-validator==1.1.3
python-dotenv==0.19.2
requests==2.32.3
